% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{annKeggEntrez}
\alias{annKeggEntrez}
\title{Get annotations for PGSEA from KEGGREST using ENTREZID.}
\usage{
annKeggEntrez(organism = "mmu", annPackage = "org.Mm.eg.db")
}
\arguments{
\item{organism}{Character organism abbreviation, default "mmu"}

\item{annPackage}{Package name wiht annotations, default "org.Mm.eg.db"}
}
\value{
A data frame with annotations and collapsed gene Symbols and ENTREZIDs
}
\description{
Get annotations for PGSEA from KEGGREST using ENTREZID.
}
