% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{boxplotRLE}
\alias{boxplotRLE}
\title{Boxplot RLE (Relative Log Expression) and correlate to RIN}
\usage{
boxplotRLE(
  expLog2,
  filePath = NULL,
  RIN = NULL,
  width = 7,
  height = 7,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{expLog2}{Expression matrix with genes in rows and samples in columns;
columns must be named}

\item{filePath}{If given, output a PDF}

\item{RIN}{If given, correlate it to median and IQR of RLE;
otherwise correlate RLE to a straight line (RLE==1 for all samples)}

\item{width}{PDF width}

\item{height}{PDF height}

\item{alpha}{Transparency of points, passed to geom_point}

\item{...}{Passed to geom_boxplot}
}
\value{
ggplot2 object and PDF if filePath is given
}
\description{
Correlate RLE to median and IQR of RIN.
If RIN is not giver, correlate it to straight line, i.e. test if RLEs are equal
}
