% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{dechiperMeltDNAtime}
\alias{dechiperMeltDNAtime}
\title{Calculate melting temperatures from ranges using a specified genome.}
\usage{
dechiperMeltDNAtime(myGRanges, myBSgenome, Trange = seq(50, 100, 1))
}
\arguments{
\item{myGRanges}{GRanges, ranges of sequences with a specified genome}

\item{myBSgenome}{BSgenome object}

\item{Trange}{Range of melting temperatures for calculation of Theta derivatives, optional, default 50:100}
}
\value{
Tibble with names of ranges (name) and Tm; additionally, sequences (seq), (max)ThetaDerivative, list of Tm (TmList); message execution time
}
\description{
Tm is estimated from a range of temperatures (Trange) by finding a max of ThetaDerivative over that range.
Tm is set to NA in case of multiple Tm values.
}
