% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{esetAnnSEs}
\alias{esetAnnSEs}
\title{Add collapsed annotations to eset using columns SYMBOLSs, ENTREZIDs}
\usage{
esetAnnSEs(eset, annDb = clariomsmousetranscriptcluster.db)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
eset with collapsed annotations in slot annotations(eset)
}
\description{
Keep probes w/o annotations
}
