% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\docType{package}
\name{FunGenPipe}
\alias{FunGenPipe}
\title{Functional genomics functions and pipes}
\description{
Tools for building function genomics pipelines
}
\section{Genomics functions that operate with ranges and sequences}{

dechiperMeltDNAtime
}

\section{Tagets meta-data}{

Functions for manipulating colors in targets.
getColPats - depricated
getColPats2
}

\section{Plot distributions, heatmap, PCA & MDS using targets meta-data}{

Note: the order od data columns should (always) be kept in the order of targets.
plotDistrTargets2
pheatmapTargets2
plotPCAtargets2
plotPCAtargets - depricated
plotMDStargets2
plotMDStargets2_MASS - depricated
}

\section{Plot distributions using limma-style data}{

plotDistr_RGList
boxplotRLE
}

\section{DEG}{
 annotations, expression (from KBlag_doxy_Cla.R via Osijek_HFHSD.R for EKocar_fibIT.R)
annotate_SEcollapsed
writeExprsGEO
}

\section{DEG & GSEA}{
 write (finalized for EKocar_fibIT.R))
getWriteHeatmap_probesTTcontrasts
getWriteHeatmap_probesTTcomparisons
getWriteHeatmap_PgseaTTcontrasts
getWriteHeatmap_PgseaTTcomparisons
}

\section{GSEA}{
 annotations & analysis
getAnnKeggEntrez
gscKeggEntrez
annTFbyFA - depricated, replaced by getAnnTransfac
getAnnTransfacEntrez
gscTransfacEntrez
}

\author{
Peter Juvan, \email{peter.juvan@gmail.com}
}
