% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getAnnKeggEntrez}
\alias{getAnnKeggEntrez}
\title{Get collapsed annotations for PGSEA from KEGGREST using ENTREZID}
\usage{
getAnnKeggEntrez(organism, annPackage, addURL = TRUE)
}
\arguments{
\item{organism}{Character KEGG organism abbreviation, e.g., "mmu", "hsa"}

\item{annPackage}{Character package name with annotations, e.g., "org.Mm.eg.db", "org.Hs.eg.db"}

\item{addURL}{Logical add a column with factor URL (factorURL), default FALSE}
}
\value{
A data frame with annotations: "KeggID", "keggPathName", collapsed gene "SYMBOLs" and "ENTREZIDs"
}
\description{
TODO: Consider adding homologous annotations
}
