% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getAnnTransfacEntrez}
\alias{getAnnTransfacEntrez}
\title{Get annotations for PGSEA from TRANSFAC using factor names (factor.FA) or factor accession (factor.AC)}
\usage{
getAnnTransfacEntrez(
  fac2drDF,
  by = "factor.FA",
  eset = NULL,
  patternExcludeFactor = NULL,
  addURL = FALSE
)
}
\arguments{
\item{fac2drDF}{Data frame TRANSFAC database}

\item{by}{Character column name from \code{fac2drDF} that is used to collapse anotations by; must start with "factor.", default "factor.FA";
annotations are not collapsed if equal to "factor.AC", but column names of the output are still in plural, e.g. "factorFAs"}

\item{eset}{Expression set with anotations ENTREZID(s) (my be collapsed by ";") within fData();
these are used only to limit TFs to ENTREZIDs; default NULL}

\item{patternExcludeFactor}{Regular expression pattern for exclusion of TFs based on column \code{by}; default NULL"}

\item{addURL}{Logical add a column with factor URL (factorURL), default FALSE}
}
\value{
A data frame with collapsed annotations including factor accessions (factorACs) and names (factorFAs), site accessions (siteACs),
gene accessions (geneACs) and short gene terms/symbols (geneSDs), and optional factor URLs (factorURLs)
}
\description{
Import TRANSFAC data frame \code{fac2drDF} and filter using ENTREZID(s) (may be collapsed by ";") from an ExpressionSet \code{eset}.
Data frame includes columns: factor.AC site.AC factor.BSq gene.AC gene.DRac2 gene.DRdbName2 factorURL factor.FA gene.SD gene.DE
TFs may be excluded based on a pattern matching their names, for example:
\code{"[^(hsa)]-miR"} will exclude non-human microRNAs
\code{"(-miR)|(isoform)"} will exclude all microRNAs and isoforms
}
\details{
Warning: annotations are unique and individually sorted; thus, they are not matched by their position within the collapsed strings.
Warning: Factor URLs will only be accessible from IPs that are licensed with QIAGEN portal \url{https://apps.ingenuity.com/ingsso/}.
Old TODO: consider creating new factors for fac2drDF_entrez_sym or even more general, e.g. using fac2drDF or  or fac2drDF_osXX.
}
