% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getColPats}
\alias{getColPats}
\title{Get color patterns for factors from Biobase::pData(dataRaw/eset) (i.e. targets)
Factors and colors share suffix in their name
Colors are prefixed with "color_", e.g. "color_Sex"}
\usage{
getColPats(featureSet)
}
\description{
Get color patterns for factors from Biobase::pData(dataRaw/eset) (i.e. targets)
Factors and colors share suffix in their name
Colors are prefixed with "color_", e.g. "color_Sex"
}
