% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getWriteHeatmap_probesTTcontrasts}
\alias{getWriteHeatmap_probesTTcontrasts}
\alias{getWriteHeatmap_probesTTcomparisons}
\alias{getWriteHeatmap_PgseaTTcontrasts}
\alias{getWriteHeatmap_PgseaTTcomparisons}
\title{Write DE genes from contrasts, return DE genes and plot heatmaps}
\usage{
getWriteHeatmap_probesTTcontrasts(
  outPath,
  esets,
  fits,
  contMatrices,
  pVals,
  varColSideColors,
  targetsOrder = NULL,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)

getWriteHeatmap_probesTTcomparisons(
  outPath,
  esets,
  fits,
  comparisons,
  pVals,
  varColSideColors,
  targetsOrder = NULL,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)

getWriteHeatmap_PgseaTTcontrasts(
  outPath,
  gscPGSEA,
  esetsPGSEA,
  fitsPGSEA,
  contMatrices,
  setIDCol,
  useNameCol,
  fitsProbes,
  esetsProbes,
  pVals,
  varColSideColors,
  targetsOrder = NULL,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)

getWriteHeatmap_PgseaTTcomparisons(
  outPath,
  gscPGSEA,
  esetsPGSEA,
  fitsPGSEA,
  comparisons,
  setIDCol,
  useNameCol,
  fitsProbes,
  esetsProbes,
  pVals,
  varColSideColors,
  targetsOrder = NULL,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)
}
\arguments{
\item{outPath}{Path to write TTs}

\item{esets}{Named list of esets for BiocGenerics::annotation of probes with columns PROBEID, SYMBOL, ENTREZID, HSA_ENTREZID}

\item{fits}{Named list of PGSEA fits}

\item{contMatrices}{Named list of contrast matrices with Levels & Contrasts}

\item{pVals}{Numeric vector p-values for heatmaps}

\item{varColSideColors}{Literal variable name from pData(esets[\link{.}]) (i.e. targets) with an associated "color_<varColSideColors>"}

\item{targetsOrder}{Order of samples, used for plotting heatmaps w/o clustering of samples; default NA}

\item{pValDE}{Numeric p-value for DEG, default 0.05}

\item{maxHeatMapProbes}{Numeric max number of DEG for heatmaps, default 50}

\item{comparisons}{Named list of comparisons made from names of contrasts}

\item{gscPGSEA}{Named list of gene set collection}

\item{esetsPGSEA}{Named list of PGSEA esets}

\item{fitsPGSEA}{Named list of PGSEA fits}

\item{setIDCol}{Add a column to TT with set IDs and name it setIDCol}

\item{fitsProbes}{Named list of probe fits}

\item{esetsProbes}{Named list of esets for BiocGenerics::annotation of probes with columns PROBEID, SYMBOL, ENTREZID, HSA_ENTREZID}

\item{setName}{Name of the column from Biobase::fData(esetsPGSEA[\link{?}]) to show with heatmaps}
}
\value{
Named list of \code{limma::topTable()} results (TTs) of DE genes / enriched sets; names corresponds to names of fits
}
\description{
File names correpond to names from fits dot(.) names from contrasts
}
\section{Functions}{
\itemize{
\item \code{getWriteHeatmap_probesTTcontrasts}: Write DE genes from contrasts, return DE genes and plot heatmaps

\item \code{getWriteHeatmap_probesTTcomparisons}: Write DE genes from comparisons, return DE genes and plot heatmaps

\item \code{getWriteHeatmap_PgseaTTcontrasts}: Write enriched pathways from contrasts, return patways and plot heatmaps

\item \code{getWriteHeatmap_PgseaTTcomparisons}: Write enriched pathways from comparisons, return patways and plot heatmaps
}}

\examples{
\dontrun{
getWriteHeatmap_PgseaTTcontrasts(outPath=file.path(resultDirOut, "3.PGSEA.KEGGREST.limma-contrasts"), gscPGSEA=gscKeggrestHsa, esetsPGSEA=esetsKeggrest, fitsPGSEA=fitsKeggrest,
                                     setIDCol="KeggID", useNameCol="keggPathNameHsa2", fitsProbes=fits, esetsProbes=esets, pVals)
getWriteHeatmap_PgseaTTcontrasts(file.path(resultDirOut, "3.PGSEA.TRANSFAC.2016.1.byFA.limma-contrasts"), gscTF.facFA, esetsTF.facFA, fitsTF.facFA,
                                         setIDCol="facFA", useNameCol="factorFA", fits, esets, pVals)
}
\dontrun{
writeHeatmap_PgseaTTcomparisons(outPath=file.path(resultDirOut, "3.PGSEA.KEGGREST.limma-comparisons"), gscPGSEA=gscKeggrestHsa, esetsPGSEA=esetsKeggrest, fitsPGSEA=fitsKeggrest,
                          setIDCol="KeggID", useNameCol="keggPathNameHsa2", fitsProbes=fits, esetsProbes=esets, pVals, targetsOrder)
writeHeatmap_PgseaTTcomparisons(file.path(resultDirOut, "3.PGSEA.TRANSFAC.2016.1.byFA.limma-comparisons"), gscTF.facFA, esetsTF.facFA, fitsTF.facFA,
                          setIDCol="facFA", useNameCol="factorFA", fits, esets, pVals, targetsOrder)
}
}
