% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getWriteHeatmap_PgseaTTcontrasts}
\alias{getWriteHeatmap_PgseaTTcontrasts}
\alias{getWriteHeatmap_PgseaTTcomparisons}
\title{Write enriched pathways from contrasts & return patways}
\usage{
getWriteHeatmap_PgseaTTcontrasts(
  outPath,
  gscPGSEA,
  esetsPGSEA,
  fitsPGSEA,
  contMatrices,
  setIDCol,
  useNameCol,
  fitsProbes,
  esetsProbes,
  pVals,
  targetsOrder = NA,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)

getWriteHeatmap_PgseaTTcomparisons(
  outPath,
  gscPGSEA,
  esetsPGSEA,
  fitsPGSEA,
  comparisons,
  setIDCol,
  useNameCol,
  fitsProbes,
  esetsProbes,
  pVals,
  targetsOrder = NA,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)
}
\arguments{
\item{outPath}{Path to write TTS}

\item{gscPGSEA}{Named list of gene set collection}

\item{esetsPGSEA}{Named list of PGSEA esets}

\item{fitsPGSEA}{Named list of PGSEA fits}

\item{contMatrices}{Named list of contrast matrices with Levels & Contrasts}

\item{setIDCol}{Add a column to TT with set IDs and name it setIDCol}

\item{fitsProbes}{Named list of probe fits}

\item{esetsProbes}{Named list of esets for BiocGenerics::annotation of probes with columns PROBEID, SYMBOL, ENTREZID, HSA_ENTREZID}

\item{comparisons}{Named list of comparisons made from names of contrasts}

\item{setName}{Name of the column from Biobase::fData(esetsPGSEA[\link{?}]) to show with heatmaps}
}
\value{
Named list of \code{limma::topTable()} results (TTs) of enriched sets; names corresponds to names of fits
}
\description{
Write enriched pathways from contrasts & return patways

Write enriched patways from comparisons & return patways
}
\examples{
\dontrun{
getWriteHeatmap_PgseaTTcontrasts(outPath=file.path(resultDirOut, "3.PGSEA.KEGGREST.limma-contrasts"), gscPGSEA=gscKeggrestHsa, esetsPGSEA=esetsKeggrest, fitsPGSEA=fitsKeggrest,
                                     setIDCol="KeggID", useNameCol="keggPathNameHsa2", fitsProbes=fits, esetsProbes=esets, pVals)
getWriteHeatmap_PgseaTTcontrasts(file.path(resultDirOut, "3.PGSEA.TRANSFAC.2016.1.byFA.limma-contrasts"), gscTF.facFA, esetsTF.facFA, fitsTF.facFA,
                                         setIDCol="facFA", useNameCol="factorFA", fits, esets, pVals)
}
\dontrun{
writeHeatmap_PgseaTTcomparisons(outPath=file.path(resultDirOut, "3.PGSEA.KEGGREST.limma-comparisons"), gscPGSEA=gscKeggrestHsa, esetsPGSEA=esetsKeggrest, fitsPGSEA=fitsKeggrest,
                          setIDCol="KeggID", useNameCol="keggPathNameHsa2", fitsProbes=fits, esetsProbes=esets, pVals, targetsOrder)
writeHeatmap_PgseaTTcomparisons(file.path(resultDirOut, "3.PGSEA.TRANSFAC.2016.1.byFA.limma-comparisons"), gscTF.facFA, esetsTF.facFA, fitsTF.facFA,
                          setIDCol="facFA", useNameCol="factorFA", fits, esets, pVals, targetsOrder)
}
}
