% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{getWriteHeatmap_probesTTcontrasts}
\alias{getWriteHeatmap_probesTTcontrasts}
\alias{getWriteHeatmap_probesTTcomparisons}
\title{Write DE genes from contrasts & return DE genes}
\usage{
getWriteHeatmap_probesTTcontrasts(
  outPath,
  esets,
  fits,
  contMatrices,
  pVals,
  targetsOrder = NA,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)

getWriteHeatmap_probesTTcomparisons(
  outPath,
  esets,
  fits,
  comparisons,
  pVals,
  targetsOrder = NA,
  pValDE = 0.05,
  maxHeatMapProbes = 50
)
}
\arguments{
\item{outPath}{Path to write TTs}

\item{esets}{Named list of esets for BiocGenerics::annotation of probes with columns PROBEID, SYMBOL, ENTREZID, HSA_ENTREZID}

\item{fits}{Named list of PGSEA fits}

\item{contMatrices}{Named list of contrast matrices with Levels & Contrasts}

\item{comparisons}{Named list of comparisons made from names of contrasts}
}
\value{
Named list of \code{limma::topTable()} results (TTs) of DE genes; names corresponds to names of fits
}
\description{
Write DE genes from contrasts & return DE genes

Write DE genes from comparisons & return DE genes
}
