% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{gscKeggEntrez}
\alias{gscKeggEntrez}
\alias{gscTransfacEntrez}
\title{Create a GeneSetCollection from probe annotations using ENTREZID}
\usage{
gscKeggEntrez(annPrb, idName = "PROBEID", organism = "mmu")

gscTransfacEntrez(
  annPrb,
  idName = "PROBEID",
  fac2drDF,
  by = "factor.FA",
  annTransfac
)
}
\arguments{
\item{annPrb}{Data frame probe annotations for mapping ENTREZID -> PROBEID; ENTREZIDs may be collapsed using ";"}

\item{idName}{Character column name from \code{annPrb} that represents IDs of probes which are used for mapping paths to a vector
of probes; default "PROBEID"}

\item{organism}{Character KEGG organism 3-letter code, e.g. "mmu", "hsa", etc.}

\item{fac2drDF}{Data frame TRANSFAC database}

\item{by}{Character column name from \code{fac2drDF} that is used to collapse anotations by; must start with "factor.", default "factor.FA";
annotations are not collapsed if equal to "factor.AC", but column names of the output are still in plural, e.g. "factorFAs"}

\item{annTransfac}{Data frame TRANSFAC annotations from \code{getAnnTransfacEntrez()}}
}
\value{
GeneSetCollection
}
\description{
Create a GeneSetCollection from probe annotations using ENTREZID
}
\section{Functions}{
\itemize{
\item \code{gscKeggEntrez}: Create a KEGG GeneSetCollection from probe annotations using ENTREZID using KEGG organism 3-letter code

\item \code{gscTransfacEntrez}: Create a TRANSFAC GeneSetCollection from probe annotations using ENTREZID
}}

