% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{gscKeggEntrez}
\alias{gscKeggEntrez}
\title{Create a list of GeneSetCollection for a list of designs and probe annotations using ENTREZID.}
\usage{
gscKeggEntrez(organism = "mmu", annPrb, annKeggEntrez, keggPathOrgIDs, design)
}
\value{
GeneSetCollection
}
\description{
Create a list of GeneSetCollection for a list of designs and probe annotations using ENTREZID.
}
