% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{pheatmapTargets}
\alias{pheatmapTargets}
\title{Plot heatmap(s) of gene expression using different distance calculation methods.}
\usage{
pheatmapTargets(
  expLog2,
  targets,
  methods = c("manhattan", "euclidean"),
  colorsFrom = "color_",
  namesFrom = NULL,
  filePath = NULL,
  width = 7,
  height = 7,
  treeheight_row = 0,
  ...
)
}
\arguments{
\item{expLog2}{Expression matrix with genes in rows and samples in columns;
columns must be named}

\item{targets}{Phenotype data with columns with colors, i.e. color_Sex}

\item{methods}{List of methods for distance calulation, individually passed to stats::dist(..., method)}

\item{colorsFrom}{String passed to getColPats2}

\item{namesFrom}{Variable passed to getColPats2}

\item{filePath}{If given, output a PDF}

\item{width}{PDF width}

\item{height}{PDF height}

\item{treeheight_row}{Passed to pheatmap; default 0; 50 for pheatmap}

\item{...}{Passed to pheatmap}
}
\value{
ggplot2 object and PDF if filePath is given
}
\description{
Plot heatmap(s) of gene expression using different distance calculation methods.
}
