% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{plotDistr_RGList}
\alias{plotDistr_RGList}
\title{Plot distributions of signals in a RGList object using multiple colors for groups of samples}
\usage{
plotDistr_RGList(
  RGList,
  filePath = NULL,
  channels = c("R", "G"),
  FUNS = c(geom_density, geom_boxplot),
  probeTypeVec = NULL,
  probeTypeValue = 0,
  numProbes = NULL,
  sampleColors = NULL,
  orderByColors = TRUE,
  scale_x_limits = NULL,
  width = 16/9 * 7,
  height = 7,
  ...
)
}
\arguments{
\item{RGList}{A list of matrices of signal intensities per channel}

\item{filePath}{NULL or character; if given, output a PDF; default NULL}

\item{channels}{List of data.frames from RGList, e.g. list("log2(R)", "log2(G)")}

\item{FUNS}{geom_functions from ggplot2, default \code{c(geom_density, geom_boxplot)}, others: \code{c(geom_violin, geom_histogram, ggridges::geom_density_ridges)}}

\item{probeTypeVec}{Vector of probe types}

\item{probeTypeValue}{Value from probeTypeVec to use for plotting}

\item{numProbes}{Number of randomly select probes}

\item{sampleColors}{Tibble of named factors of colors with names matching sample names}

\item{orderByColors}{Logical for ordering samples by colors (coded as a factor with levels in order); default TRUE}

\item{scale_x_limits}{NULL for auto-scale; use c(0,16) or less for log2(intensities)}

\item{width}{PDF width, default 16/9*7=12.44}

\item{height}{PDF height, default 7}

\item{...}{Passed to ggplot2::FUN, e.g.: bins, binwidth, show.legend}
}
\value{
A list of ggplots, one per a combination of FUNS, channels and sampleColors
PDF if filePath is not NULL
}
\description{
Plots distributions for each combination of FUNS, channels and sampleColors.
By default, samples are ordered according to the sampleColors (if provided).
Colors should be provided as factors with a preset order of levels; the order of colors is determined by the order of levels
}
\section{Implementation}{

See \url{https://www.tidyverse.org/blog/2020/02/glue-strings-and-tidy-eval/} for bracing variables.
Evaluation of expression: \url{https://adv-r.hadley.nz/evaluation.html}.
For \code{bquote(a +. (b))}: \url{http://adv-r.had.co.nz/Expressions.html}.
}

\examples{
\dontrun{
plotDistr_RGList(dataRG.bgc, file.path(curDir, "density-boxplot_dataRG.bgc_genes.pdf"), 
channels = c("log2R","log2G","beta"), probeTypeVec = dataRG.bgc$genes$ControlType, probeTypeValue = 0,
numProbes = 10000, sampleColors = targets \%>\% select(starts_with("color_")))
## Removed parameters/code:
# transform = log2: transformation function applied to each channel (default: log2; identity for none)
mutate("{{transform}}({ch})" := transform(!!parse_expr(ch)))
ggplot(aes(x=transform(!!parse_expr(ch)), color=HybName)) +
ggplot(d2, aes(x=!!parse_expr(ch), color=fct_reorder(HybName, order(colors)))) + #, order=colors)) + 
}
}
