% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{plotMDStargets2}
\alias{plotMDStargets2}
\alias{plotMDStargets2_MASS}
\title{Plot MDS(es) of gene expression using different distance calculation methods and alternative colors.}
\usage{
plotMDStargets2(
  expLog2,
  targets,
  colorsFrom = "color_",
  rename = TRUE,
  namesFrom = NULL,
  scale = FALSE,
  center = FALSE,
  FUNS = c("cmdscale", "isoMDS", "sammon"),
  p = 2,
  selection = "pairwise",
  tops = 500,
  maxit = 50,
  trace = FALSE,
  tol = 0.001,
  size = 3,
  filePath = NULL,
  width = 7,
  height = 7,
  ...
)

plotMDStargets2_MASS(
  expLog2,
  targets,
  colorsFrom = "color_",
  namesFrom = NULL,
  scale = FALSE,
  center = FALSE,
  methods = c("euclidean", "manhattan"),
  FUNS = c("isoMDS", "sammon"),
  p = 2,
  maxit = 50,
  trace = FALSE,
  tol = 0.001,
  size = 3,
  filePath = NULL,
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{expLog2}{Expression matrix (preferably in log2 scale) with genes in rows and samples in columns}

\item{targets}{Phenotype data with columns with colors, e.g., color_Sex}

\item{colorsFrom}{String variable prefix(es) from targets, passed to \code{getColPats2()}; default "color_"}

\item{rename}{Logical, default TRUE, passed to \code{getColPats2()}; rename colors using associated variables
or a variable namesFrom (default)}

\item{namesFrom}{Variable from targets, passed to \code{getColPats2()}; suggested HybName}

\item{scale}{Logical scale data, passed to \code{MASS_MDScols()}}

\item{center}{Logical center data, passed to \code{MASS_MDScols()}}

\item{FUNS}{List of MDS function from pacakge MASS, passed individually to \code{MASS_MDScols()}}

\item{p}{Power of the Minkowski distance, passed to \code{dist()} through \code{MASS_MDScols()}}

\item{selection}{Character "pairwise" or "common" for selection of rows or NULL for using all rows; default "pairwise"}

\item{tops}{Integer vector number of rows (genes), passed individually to \code{MDScols()}}

\item{maxit}{Passed to \code{MASS_MDScols()},}

\item{trace}{Trace progress, passed to \code{MASS_MDScols()}, default FALSE}

\item{tol}{Tolerance, passed to \code{MASS_MDScols()}}

\item{size}{Size of ggplot labels, passed to \code{geom_text()}}

\item{filePath}{NULL or character; if given, output a PDF; default NULL}

\item{width}{PDF width}

\item{height}{PDF height}

\item{...}{Passed to ggplot2::geom_text}

\item{methods}{List of methods for distance calulation, passed individually to \code{stats::dist(method)} through \code{MASS_MDScols()}}
}
\value{
Invisibly a list of ggplot2 objects and PDF if filePath is given

Invisibly a list of ggplot2 objects and PDF if filePath is given
}
\description{
Plot MDS(es) of gene expression using different distance calculation methods and alternative colors.

Depricated: Plot MDS(es) of gene expression using different distance calculation methods and alternative colors.
}
\section{TODO}{
 add parameter dim.plot as in limma::plotMDS
}

