% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{plotPCAtargets}
\alias{plotPCAtargets}
\title{Plot PCA for gene expression and phenotype data}
\usage{
plotPCAtargets(
  expLog2,
  targets,
  shape,
  color,
  fill,
  size,
  scale_color = NULL,
  scale_fill = NULL,
  filePath = NULL,
  width = 7,
  height = 7,
  stroke = 1,
  guides_fill = "none",
  alpha = 0.5,
  ...
)
}
\arguments{
\item{expLog2}{Gene expression matrix in log2 scale}

\item{targets}{Phenotype data with columns shape, color, fill, size,
as well as color_color and color_fill}

\item{shape}{Variable from targets for ggplot2::aes}

\item{color}{Variable from targets for ggplot2::aes}

\item{fill}{Variable from targets for ggplot2::aes}

\item{size}{Variable from targets for ggplot2::aes}

\item{scale_color}{Variable from targets, named vector for scale_color_manual(value=scale_color)}

\item{scale_fill}{Variable from targets, named vector for scale_fill_manual(value=scale_fill)}

\item{filePath}{PDF}

\item{width}{PDF width}

\item{height}{PDF height}

\item{stroke}{Line thickness, passed to geom_point}

\item{guides_fill}{Default "none"; use "legend" to display it}

\item{alpha}{Transparency of points, passed to geom_point}

\item{...}{Passed to geom_point}
}
\value{
ggplot2 object and PDF if filePath is given
}
\description{
Plot PCA for gene expression and phenotype data
}
\section{TODO}{

FIX: add return value
FIX: Discrete scale_color supplied for continuous color variable; scale_color not used
Fix: case that scale_color and scale_fill are not named vectors or NULL
Fix scale_size: Error: Discrete value supplied to continuous scale
Use scale_colour_viridis_c
}

\section{Implementation}{

Put ggplot() inside {} to be able to access data using dot (.)
}

