% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{writeExprsGEO}
\alias{writeExprsGEO}
\title{Write gene expression values for GEO}
\usage{
writeExprsGEO(esetAnn, gplAccNum, filePath = NULL)
}
\arguments{
\item{esetAnn}{ExpressionSet}

\item{gplAccNum}{Character GEO GPL accession number}

\item{filePath}{Character path/fileName.txt}
}
\value{
Matrix expression values for GEO.
}
\description{
Downloads platform info file from GEO using GPL accession number \code{gplAccNum}.
Adds platform IDs that are missing from ExpressionSet \code{esetAnn} and sets their values to NA.
Writes data to tad-delimited file \code{filePath}.
Uses \code{rownames(esetAnn)} as platform IDs.
}
