% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{annKeggEntrez}
\alias{annKeggEntrez}
\title{Get annotations for PGSEA from KEGGREST using ENTREZID.}
\usage{
annKeggEntrez(organism = "mmu", annPackage = "org.Mm.eg.db")
}
\description{
Get annotations for PGSEA from KEGGREST using ENTREZID.
}
