% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{dechiperMeltDNAtime}
\alias{dechiperMeltDNAtime}
\title{Calculate melting temperatures from ranges using a specified genome.}
\usage{
dechiperMeltDNAtime(
  myGRanges,
  myBSgenome,
  Trange = seq(50, 100, 1),
  ions = 0.2,
  dropMcols = c("seq", "Trange", "ThetaDerivative", "maxThetaDerivative", "TmList")
)
}
\arguments{
\item{myGRanges}{GRanges, ranges of sequences with a specified genome}

\item{myBSgenome}{BSgenome object}

\item{Trange}{Numeric vector of melting temperatures for calculation of Theta derivatives, forwarded to \code{DECIPHER::MeltDNA}; optional, default 50:100}

\item{ions}{Numeric molar sodium equivalent ionic concentration, forwarded to \code{DECIPHER::MeltDNA}; optional, default 0.2}

\item{dropMcols}{Intermediate columns to drop, default c("seq","Trange","ThetaDerivative","maxThetaDerivative","TmList"), NULL to include all}
}
\value{
GRanges with melting temperatures added to mcols column "Tm" and optional intermediate columns with sequences (seq), (max)ThetaDerivative, list of Tm (TmList); message execution time
}
\description{
Tm is estimated from a range of temperatures (Trange) by finding a max of ThetaDerivative over that range.
Tm is set to NA in case of multiple Tm values.
}
