% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunGenPipe.R
\name{plotMDStargets2}
\alias{plotMDStargets2}
\title{Plot MDS(es) of gene expression using different distance calculation methods and alternative colors.}
\usage{
plotMDStargets2(
  expLog2,
  targets,
  colorsFrom = "color_",
  namesFrom = NULL,
  scale = FALSE,
  center = FALSE,
  methods = c("euclidean", "manhattan"),
  FUNS = c("isoMDS", "sammon"),
  p = 2,
  maxit = 50,
  trace = FALSE,
  tol = 0.001,
  size = 3,
  filePath = NULL,
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{expLog2}{Expression matrix with genes in rows and samples in columns;
columns must be named}

\item{targets}{Phenotype data with columns with colors, i.e. color_Sex}

\item{colorsFrom}{String variable prefix(es) from targets, passed to \code{getColPats2()}; default "color_"}

\item{namesFrom}{Variable from targets, passed to \code{getColPats2()}; suggested HybName}

\item{scale}{Logical scale data, passed to \code{MASS_MDScols()}}

\item{center}{Logical center data, passed to \code{MASS_MDScols()}}

\item{methods}{List of methods for distance calulation, passed individually to \code{stats::dist(method)} through \code{MASS_MDScols()}}

\item{FUNS}{List of MDS function from pacakge MASS, passed individually to \code{MASS_MDScols()}}

\item{p}{Power of the Minkowski distance, passed to \code{dist()} through \code{MASS_MDScols()}}

\item{maxit}{Passed to \code{MASS_MDScols()},}

\item{trace}{Trace progress, passed to \code{MASS_MDScols()}, default FALSE}

\item{tol}{Tolerance, passed to \code{MASS_MDScols()}}

\item{size}{Size of ggplot labels, passed to \code{geom_text()}}

\item{filePath}{If given, output a PDF}

\item{width}{PDF width}

\item{height}{PDF height}

\item{...}{Passed to ggplot2::geom_text}
}
\value{
Invisibly a list of ggplot2 objects and PDF if filePath is given
}
\description{
Plot MDS(es) of gene expression using different distance calculation methods and alternative colors.
}
